+++ src/DI/Container.php
@@ -9,30 +9,32 @@
 
 namespace DI;
 
-use DI\Definition\AliasDefinition;
 use DI\Definition\ClassDefinition;
-use DI\Definition\CallableDefinition;
+use DI\Definition\Definition;
 use DI\Definition\DefinitionManager;
 use DI\Definition\ValueDefinition;
 use DI\DefinitionHelper\DefinitionHelper;
+use DI\DefinitionResolver\AliasDefinitionResolver;
+use DI\DefinitionResolver\CallableDefinitionResolver;
+use DI\DefinitionResolver\ClassDefinitionResolver;
+use DI\DefinitionResolver\DefinitionResolver;
+use DI\DefinitionResolver\ValueDefinitionResolver;
 use Exception;
 use InvalidArgumentException;
-use ProxyManager\Configuration as ProxyManagerConfiguration;
 use ProxyManager\Factory\LazyLoadingValueHolderFactory;
-use ProxyManager\GeneratorStrategy\EvaluatingGeneratorStrategy;
 
 /**
- * Dependency Injection Container
+ * Dependency Injection Container.
  *
  * @author Matthieu Napoli <matthieu@mnapoli.fr>
  */
-class Container
+class Container implements ContainerInterface
 {
     /**
-     * Map of instances of entry with Singleton scope
+     * Map of entries with Singleton scope
      * @var array
      */
-    private $entries = array();
+    private $resolvedEntries = array();
 
     /**
      * @var DefinitionManager
@@ -40,103 +42,99 @@ class Container
     private $definitionManager;
 
     /**
-     * @var Injector
+     * Map of definition resolvers, indexed by the classname of the definition it resolves.
+     *
+     * @var DefinitionResolver[]
      */
-    private $injector;
+    private $definitionResolvers;
 
     /**
-     * @var LazyLoadingValueHolderFactory
-     */
-    private $proxyFactory;
-
-    /**
-     * Array of classes being instantiated.
-     * Used to avoid circular dependencies.
+     * Array of entries being resolved. Used to avoid circular dependencies and infinite loops.
      * @var array
      */
-    private $classesBeingInstantiated = array();
+    private $entriesBeingResolved = array();
 
     /**
-     * Parameters are optional, use them to override a dependency.
+     * Use the ContainerBuilder to ease constructing the Container.
      *
-     * @param DefinitionManager|null             $definitionManager
-     * @param Injector|null              $injector
-     * @param LazyLoadingValueHolderFactory|null $proxyFactory
+     * @see ContainerBuilder
+     *
+     * @param DefinitionManager             $definitionManager
+     * @param LazyLoadingValueHolderFactory $proxyFactory
+     * @param ContainerInterface            $wrapperContainer  If the container is wrapped by another container.
      */
     public function __construct(
-        DefinitionManager $definitionManager = null,
-        Injector $injector = null,
-        LazyLoadingValueHolderFactory $proxyFactory = null
+        DefinitionManager $definitionManager,
+        LazyLoadingValueHolderFactory $proxyFactory,
+        ContainerInterface $wrapperContainer = null
     ) {
-        $this->definitionManager = $definitionManager ?: new DefinitionManager(true, true);
-        $this->injector = $injector ?: new DefaultInjector($this);
-        $this->proxyFactory = $proxyFactory ?: $this->createDefaultProxyFactory();
+        $this->definitionManager = $definitionManager;
+
+        // Definition resolvers
+        $wrapperContainer = $wrapperContainer ?: $this;
+        $this->definitionResolvers = array(
+            'DI\Definition\ValueDefinition'    => new ValueDefinitionResolver(),
+            'DI\Definition\CallableDefinition' => new CallableDefinitionResolver($wrapperContainer),
+            'DI\Definition\AliasDefinition'    => new AliasDefinitionResolver($wrapperContainer),
+            'DI\Definition\ClassDefinition'    => new ClassDefinitionResolver($wrapperContainer, $proxyFactory),
+        );
 
         // Auto-register the container
-        $this->entries[get_class($this)] = $this;
+        $this->resolvedEntries[get_class($this)] = $this;
     }
 
     /**
-     * Returns an instance by its name
+     * Returns an entry of the container by its name.
      *
-     * @param string $name Entry name or a class name
-     * @param bool   $useProxy If true, returns a proxy class of the instance
-     *                         if it is not already loaded
-     * @throws InvalidArgumentException
-     * @throws DependencyException
-     * @throws NotFoundException
-     * @return mixed Instance
+     * @param string $name Entry name or a class name.
+     *
+     * @throws InvalidArgumentException The name parameter must be of type string.
+     * @throws DependencyException Error while resolving the entry.
+     * @throws NotFoundException No entry found for the given name.
+     * @return mixed
      */
-    public function get($name, $useProxy = false)
+    public function get($name)
     {
-        if (!is_string($name)) {
+        if (! is_string($name)) {
             throw new InvalidArgumentException("The name parameter must be of type string");
         }
 
         // Try to find the entry in the map
-        if (array_key_exists($name, $this->entries)) {
-            return $this->entries[$name];
+        if (array_key_exists($name, $this->resolvedEntries)) {
+            return $this->resolvedEntries[$name];
         }
 
         // Entry not loaded, use the definitions
         $definition = $this->definitionManager->getDefinition($name);
 
-        // It's a value
-        if ($definition instanceof ValueDefinition) {
-            $this->entries[$name] = $definition->getValue();
-            return $this->entries[$name];
+        if (! $definition) {
+            throw new NotFoundException("No entry or class found for '$name'");
         }
 
-        // It's a closure
-        if ($definition instanceof CallableDefinition) {
-            $callable = $definition->getCallable();
-            $this->entries[$name] = $callable($this);
-            return $this->entries[$name];
+        $definitionResolver = $this->getDefinitionResolver($definition);
+
+        // Check if we are already getting this entry -> circular dependency
+        if (isset($this->entriesBeingResolved[$name])) {
+            throw new DependencyException("Circular dependency detected while trying to get entry '$name'");
+        }
+        $this->entriesBeingResolved[$name] = true;
+
+        // Resolve the definition
+        try {
+            $value = $definitionResolver->resolve($definition);
+        } catch (Exception $exception) {
+            unset($this->entriesBeingResolved[$name]);
+            throw $exception;
         }
 
-        // It's an alias
-        if ($definition instanceof AliasDefinition) {
-            return $this->get($definition->getTargetEntryName(), $useProxy);
+        unset($this->entriesBeingResolved[$name]);
+
+        // If the entry is singleton, we store it to always return it without recomputing it.
+        if ($definition->getScope() == Scope::SINGLETON()) {
+            $this->resolvedEntries[$name] = $value;
         }
 
-        // It's a class
-        if ($definition instanceof ClassDefinition) {
-            // Return a proxy class
-            if ($useProxy || $definition->isLazy()) {
-                $instance = $this->getProxy($definition);
-            } else {
-                $instance = $this->getNewInstance($definition);
-            }
-
-            if ($definition->getScope() == Scope::SINGLETON()) {
-                // If it's a singleton, store the newly created instance
-                $this->entries[$name] = $instance;
-            }
-
-            return $instance;
-        }
-
-        throw new NotFoundException("No entry or class found for '$name'");
+        return $value;
     }
 
     /**
@@ -152,7 +150,7 @@ class Container
             throw new InvalidArgumentException("The name parameter must be of type string");
         }
 
-        return array_key_exists($name, $this->entries) || $this->definitionManager->getDefinition($name);
+        return array_key_exists($name, $this->resolvedEntries) || $this->definitionManager->getDefinition($name);
     }
 
     /**
@@ -160,16 +158,17 @@ class Container
      *
      * @param object $instance Object to perform injection upon
      * @throws InvalidArgumentException
-     * @throws DependencyException
+     * @throws DependencyException Error while injecting dependencies
      * @return object $instance Returns the same instance
      */
     public function injectOn($instance)
     {
         $definition = $this->definitionManager->getDefinition(get_class($instance));
+        $definitionResolver = $this->getDefinitionResolver($definition);
 
         // Check that the definition is a class definition
-        if ($definition instanceof ClassDefinition) {
-            $instance = $this->injector->injectOnInstance($definition, $instance);
+        if ($definition instanceof ClassDefinition && $definitionResolver instanceof ClassDefinitionResolver) {
+            $definitionResolver->injectOnInstance($definition, $instance);
         }
 
         return $instance;
@@ -184,8 +183,8 @@ class Container
     public function set($name, $value)
     {
         // Clear existing entry if it exists
-        if (array_key_exists($name, $this->entries)) {
-            unset($this->entries[$name]);
+        if (array_key_exists($name, $this->resolvedEntries)) {
+            unset($this->resolvedEntries[$name]);
         }
 
         if ($value instanceof DefinitionHelper) {
@@ -208,38 +207,6 @@ class Container
     }
 
     /**
-     * @param Injector $injector
-     */
-    public function setInjector(Injector $injector)
-    {
-        $this->injector = $injector;
-    }
-
-    /**
-     * @return Injector
-     */
-    public function getInjector()
-    {
-        return $this->injector;
-    }
-
-    /**
-     * @return LazyLoadingValueHolderFactory
-     */
-    public function getProxyFactory()
-    {
-        return $this->proxyFactory;
-    }
-
-    /**
-     * @param LazyLoadingValueHolderFactory $proxyFactory
-     */
-    public function setProxyFactory(LazyLoadingValueHolderFactory $proxyFactory)
-    {
-        $this->proxyFactory = $proxyFactory;
-    }
-
-    /**
      * @return DefinitionManager
      */
     public function getDefinitionManager()
@@ -248,62 +215,20 @@ class Container
     }
 
     /**
-     * @param ClassDefinition $classDefinition
-     * @throws DependencyException
-     * @throws \Exception
-     * @return object The instance
-     */
-    private function getNewInstance(ClassDefinition $classDefinition)
-    {
-        $classname = $classDefinition->getClassName();
-
-        if (isset($this->classesBeingInstantiated[$classname])) {
-            throw new DependencyException("Circular dependency detected while trying to instantiate class '$classname'");
-        }
-        $this->classesBeingInstantiated[$classname] = true;
-
-        try {
-            $instance = $this->injector->createInstance($classDefinition);
-        } catch (Exception $exception) {
-            unset($this->classesBeingInstantiated[$classname]);
-            throw $exception;
-        }
-
-        unset($this->classesBeingInstantiated[$classname]);
-        return $instance;
-    }
-
-    /**
-     * Returns a proxy instance
+     * Returns a resolver capable of handling the given definition.
      *
-     * @param ClassDefinition $definition
-     * @return object Proxy instance
+     * @param Definition $definition
+     * @throws \RuntimeException No definition resolver was found for this type of definition.
+     * @return DefinitionResolver
      */
-    private function getProxy(ClassDefinition $definition)
+    private function getDefinitionResolver(Definition $definition)
     {
-        $proxy = $this->proxyFactory->createProxy(
-            $definition->getClassName(),
-            function (& $wrappedObject, $proxy, $method, $parameters, & $initializer) use ($definition) {
-                $wrappedObject = $this->getNewInstance($definition);
-                $initializer = null; // turning off further lazy initialization
-                return true;
-            }
-        );
+        $definitionType = get_class($definition);
 
-        return $proxy;
-    }
+        if (! isset($this->definitionResolvers[$definitionType])) {
+            throw new \RuntimeException("No definition resolver was configured for definition of type $definitionType");
+        }
 
-    /**
-     * @return LazyLoadingValueHolderFactory
-     */
-    private function createDefaultProxyFactory()
-    {
-        // Proxy factory
-        $config = new ProxyManagerConfiguration();
-        // By default, auto-generate proxies and don't write them to file
-        $config->setAutoGenerateProxies(true);
-        $config->setGeneratorStrategy(new EvaluatingGeneratorStrategy());
-
-        return new LazyLoadingValueHolderFactory($config);
+        return $this->definitionResolvers[$definitionType];
     }
 }