+++ src/DI/ContainerBuilder.php
@@ -17,13 +17,21 @@ use ProxyManager\Factory\LazyLoadingValueHolderFactory;
 use ProxyManager\GeneratorStrategy\EvaluatingGeneratorStrategy;
 
 /**
- * Fluent helper to create a Container
+ * Helper to create and configure a Container.
  *
- * @since 3.2
+ * With the default options, the container created is appropriate for the development environment.
+ *
+ * @since  3.2
  * @author Matthieu Napoli <matthieu@mnapoli.fr>
  */
 class ContainerBuilder
 {
+    /**
+     * Name of the container class, used to create the container.
+     * @var string
+     */
+    private $containerClass;
+
     /**
      * @var boolean
      */
@@ -58,8 +66,28 @@ class ContainerBuilder
 
     /**
      * If PHP-DI is wrapped in another container, this references the wrapper.
+     * @var ContainerInterface
      */
     private $wrapperContainer;
+
+    /**
+     * Build a container configured for the dev environment.
+     *
+     * @return Container
+     */
+    public static function buildDevContainer()
+    {
+        $builder = new self();
+        return $builder->build();
+    }
+
+    /**
+     * @param string $containerClass Name of the container class, used to create the container.
+     */
+    public function __construct($containerClass = 'DI\Container')
+    {
+        $this->containerClass = $containerClass;
+    }
 
     /**
      * @return Container
@@ -75,22 +103,12 @@ class ContainerBuilder
             $definitionManager->setCache($this->cache);
         }
 
-        // Injector
-        $injector = $this->wrapperContainer ? new DefaultInjector($this->wrapperContainer) : null;
-
         // Proxy factory
-        $config = new Configuration();
-        $config->setAutoGenerateProxies(true);
-        if ($this->writeProxiesToFile) {
-            $config->setProxiesTargetDir($this->proxyDirectory);
-            spl_autoload_register($config->getProxyAutoloader());
-        } else {
-            $config->setGeneratorStrategy(new EvaluatingGeneratorStrategy());
-        }
+        $proxyFactory = $this->buildProxyFactory();
 
-        $container = new Container($definitionManager, $injector, new LazyLoadingValueHolderFactory($config));
+        $containerClass = $this->containerClass;
 
-        return $container;
+        return new $containerClass($definitionManager, $proxyFactory, $this->wrapperContainer);
     }
 
     /**
@@ -150,7 +168,7 @@ class ContainerBuilder
      * For dev environment, use writeProxiesToFile(false) (default configuration)
      * For production environment, use writeProxiesToFile(true, 'tmp/proxies')
      *
-     * @param boolean     $writeToFile If true, write the proxies to disk to improve performances
+     * @param boolean     $writeToFile    If true, write the proxies to disk to improve performances
      * @param string|null $proxyDirectory Directory where to write the proxies
      * @return ContainerBuilder
      *
@@ -161,7 +179,9 @@ class ContainerBuilder
         $this->writeProxiesToFile = $writeToFile;
 
         if ($writeToFile && $proxyDirectory === null) {
-            throw new InvalidArgumentException("The proxy directory must be specified if you want to write proxies on disk");
+            throw new InvalidArgumentException(
+                "The proxy directory must be specified if you want to write proxies on disk"
+            );
         }
         $this->proxyDirectory = $proxyDirectory;
 
@@ -169,16 +189,34 @@ class ContainerBuilder
     }
 
     /**
-     * If PHP-DI's container is wrapped by another container (for example Acclimate), we can
+     * If PHP-DI's container is wrapped by another container, we can
      * set this so that PHP-DI will use the wrapper rather than itself for building objects.
      *
-     * @param object $otherContainer This container only need to implement the get($name) method
+     * @param ContainerInterface $otherContainer
      * @return $this
      */
-    public function wrapContainer($otherContainer)
+    public function wrapContainer(ContainerInterface $otherContainer)
     {
         $this->wrapperContainer = $otherContainer;
 
         return $this;
+    }
+
+    /**
+     * @return LazyLoadingValueHolderFactory
+     */
+    private function buildProxyFactory()
+    {
+        $config = new Configuration();
+        $config->setAutoGenerateProxies(true);
+
+        if ($this->writeProxiesToFile) {
+            $config->setProxiesTargetDir($this->proxyDirectory);
+            spl_autoload_register($config->getProxyAutoloader());
+        } else {
+            $config->setGeneratorStrategy(new EvaluatingGeneratorStrategy());
+        }
+
+        return new LazyLoadingValueHolderFactory($config);
     }
 }