+++ src/DI/Definition/ValueDefinition.php
@@ -9,6 +9,8 @@
 
 namespace DI\Definition;
 
+use DI\Scope;
+
 /**
  * Definition of a value for dependency injection
  *
@@ -44,6 +46,16 @@ class ValueDefinition implements Definition
     public function getName()
     {
         return $this->name;
+    }
+
+    /**
+     * A value definition is like a constant, there is nothing to compute, the value is the same for everyone.
+     *
+     * {@inheritdoc}
+     */
+    public function getScope()
+    {
+        return Scope::SINGLETON();
     }
 
     /**