+++ src/DI/DefinitionResolver/DefinitionResolver.php
@@ -1,0 +1,29 @@
+<?php
+/**
+ * PHP-DI
+ *
+ * @link      http://mnapoli.github.com/PHP-DI/
+ * @copyright Matthieu Napoli (http://mnapoli.fr/)
+ * @license   http://www.opensource.org/licenses/mit-license.php MIT (see the LICENSE file)
+ */
+
+namespace DI\DefinitionResolver;
+
+use DI\Definition\Definition;
+
+/**
+ * Represents an object that resolves a DI\Definition to a value.
+ *
+ * @since 4.0
+ * @author Matthieu Napoli <matthieu@mnapoli.fr>
+ */
+interface DefinitionResolver
+{
+    /**
+     * Resolve a definition to a value.
+     *
+     * @param Definition $definition Object that defines how the value should be obtained.
+     * @return mixed Value obtained from the definition.
+     */
+    public function resolve(Definition $definition);
+}