+++ tests/IntegrationTests/DI/Issues/Issue70and76Test.php
@@ -9,7 +9,7 @@
 
 namespace IntegrationTests\DI\Issues;
 
-use DI\Container;
+use DI\ContainerBuilder;
 use DI\Entry;
 
 /**
@@ -23,7 +23,7 @@ class Issue70and76Test extends \PHPUnit_Framework_TestCase
      */
     public function valueDefinitionShouldOverrideReflectionDefinition()
     {
-        $container = new Container();
+        $container = ContainerBuilder::buildDevContainer();
 
         $container->set('stdClass', 'foo');
         $this->assertEquals('foo', $container->get('stdClass'));
@@ -34,11 +34,11 @@ class Issue70and76Test extends \PHPUnit_Framework_TestCase
      */
     public function closureDefinitionShouldOverrideReflectionDefinition()
     {
-        $container = new Container();
+        $container = ContainerBuilder::buildDevContainer();
 
-        $container->set('stdClass', Entry::factory(function() {
-                return 'foo';
-            }));
+        $container->set('stdClass', Entry::factory(function () {
+            return 'foo';
+        }));
         $this->assertEquals('foo', $container->get('stdClass'));
     }
 }