+++ tests/UnitTests/DI/Definition/CallableDefinitionTest.php
@@ -10,6 +10,7 @@
 namespace UnitTests\DI\Definition;
 
 use DI\Definition\CallableDefinition;
+use DI\Scope;
 
 /**
  * Test class for CallableDefinition
@@ -19,12 +20,13 @@ class CallableDefinitionTest extends \PHPUnit_Framework_TestCase
     public function testGetters()
     {
         $callable = function () {
-            return 1;
         };
         $definition = new CallableDefinition('foo', $callable);
 
         $this->assertEquals('foo', $definition->getName());
         $this->assertEquals($callable, $definition->getCallable());
+        // Default scope
+        $this->assertEquals(Scope::SINGLETON(), $definition->getScope());
     }
 
     /**
@@ -37,6 +39,14 @@ class CallableDefinitionTest extends \PHPUnit_Framework_TestCase
 
         $this->assertEquals('foo', $definition->getName());
         $this->assertEquals($callable, $definition->getCallable());
+    }
+
+    public function testScope()
+    {
+        $definition = new CallableDefinition('foo', function () {
+        }, Scope::PROTOTYPE());
+
+        $this->assertEquals(Scope::PROTOTYPE(), $definition->getScope());
     }
 
     public function testMergeable()