+++ change-log.md
@@ -2,10 +2,16 @@
 
 ## 4.0
 
+Major changes:
+
+* The configuration format has changed
+
 BC-breaks:
 
 * XML and JSON definitions have been removed
 * `ContainerSingleton` has been removed
+* You cannot configure an injection as lazy anymore, you can only configure a container entry as lazy
+* The Container constructor now takes mandatory parameters. Use the ContainerBuilder to create a Container.
 
 All changes:
 
@@ -15,7 +21,7 @@ All changes:
 * [#110](https://github.com/mnapoli/PHP-DI/issues/110) XML definitions are not supported anymore
 * [#122](https://github.com/mnapoli/PHP-DI/issues/122) JSON definitions are not supported anymore
 * `ContainerSingleton` has finally been removed
-* `FactoryInterface` has been renamed to `Injector`, only people implementing their own factory/injector should care
+* Added `ContainerBuilder::buildDevContainer()` to get started with a default container very easily.
 
 ## 3.5.1